Added 404 Error,About,Service page & change navbar style
