+++ src/components/User/About.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const About = () => {
+    return (
+        <>
+            <hr />
+            <h3 className="text-center ">About Page</h3>
+            <hr />
+        </>
+    );
+};
+
+export default About;
