+++ src/components/User/Error Page/Error.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+export const Error = () => {
+    return (
+        <>
+            <hr />
+            <div className="container mt-5 text-center">
+                <h2 className="text-danger"> 404 Error Page </h2>
+                <h5 className="text-muted">Sorry, This Page does not exist</h5>
+            </div>
+            <hr />
+        </>
+    );
+};
+
+export default Error;
