+++ src/components/User/Services.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Services = () => {
+    return (
+        <>
+            <hr />
+            <h3 className="text-center ">Service Page</h3>
+            <hr />
+        </>
+    );
+};
+
+export default Services;
