+++ src/components/User/UsersData/UserComponent.css
@@ -22,7 +22,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    font-size: 39px;
+    font-size: 25px;
 }
 
 .badge span {
@@ -43,4 +43,13 @@
     background-color: white;
     border: 1px solid rgb(228, 228, 228);
     padding: 5px;
+}
+.table td,
+.table th {
+    vertical-align: middle;
+}
+*,
+::after,
+::before {
+    box-sizing: content-box;
 }
